+++ src/routes/(index)/[[lang=lang]]/[slug]/+page.svelte
@@ -45,13 +45,13 @@
 {#if featureImage }
 <meta property="og:image" content="{featureImage}" />
 
-{#if config.farcasterFrameV2Enabled}
+{#if config.farcaster.miniAppEnabled}
 <meta name="fc:frame" content={generateURLFCFrameEmbed(featureImage, pageUrl)} />
 {/if}
 {:else}
 <meta property="og:image" content={`${config.baseSiteUrl}/images/og/default-og-blog-opt.webp`} />
 
-{#if config.farcasterFrameV2Enabled}
+{#if config.farcaster.miniAppEnabled}
 <meta name="fc:frame" content={generateURLFCFrameEmbed('', pageUrl)} />
 {/if}